requirements.txt
# Core
arelle-release>=2.20.5
xlsxwriter>=3.2.0
reportlab>=4.1.0
openpyxl>=3.1.2
requests>=2.31.0
cryptography>=42.0.0

# Optional / GUI
pillow>=10.3.0

# Dev
psutil>=5.9.8

